Camera full Operational
Press V to swap view to arial/character.
In arial view, the mouse cursor is visible, you can pan the view by moving it to the edges of the screen.
You can then press C to center the viewpoint on the character.
Press V to return to character, where the mouse cursor is again hidden.